Work toward FIFO-based key/value store
This is probably on a good idea, for two reasons.
a) shsh-writeval will block if there is no reader at the other
end of the FIFO, thereby blocking the upstream processes.
b) Apparently FIFOs don't work well when there are multiple readers.

Therefore this change stays on a branch and the trunk proceeds
to implement the store using Unix and Internet domain sockets.