Avoid deadlock caused by successive cat commands
Each of the cat commands depends on its preceding one. Therefore,
the corresponding stream file is not opened, until the preceding
one finishes. This prevents the shell from opening the output
redirection to the corresponding write end of the named pipe,
and thereby never launching the corresponding command.

Solve the problem by using a single cat command, which the
d50f56d34 version of sgsh replaces with a call to sgsh-tee,
which opens all the input files upon startup.

TODO: Have sgsh automatically flag such deadlocks.