Debug concentrator; add exemplary sgsh graph in tests.
Keep sending the node array with the message block in order
to be able to specify a node's pid given its index in the array.
Count the times a block is received on a specific port
at state PS_RUN only.
Allow the concentrator to terminate after writing
to the process whose id is set.